+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Change CSW harvesters XML processor from lxml to saxonche [#3380](https://github.com/opendatateam/udata/pull/3380)
 - Convert security views to cdata [#3348](https://github.com/opendatateam/udata/pull/3348/)
   - Add `SECURITY_SPA_ON_SAME_DOMAIN` to `True` if your frontend is on the same domain as the backend.
+  - Change the redirects URLs in ProConnect:
+    - From https://www.data.gouv.fr/proconnect/auth to https://www.data.gouv.fr/api/1/proconnect/auth
+    - From https://www.data.gouv.fr/proconnect/logout to https://www.data.gouv.fr/api/1/proconnect/logout
+
 
 ## 10.9.0 (2025-08-28)